derm1m_exp/baseline/requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0

# Vision-Language Models
qwen-vl-utils
openai>=1.0.0

# Image and Video Processing
opencv-python>=4.8.0
Pillow>=9.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
tqdm>=4.65.0

# Flash Attention (optional but recommended for faster inference)
# Install separately if needed: pip install flash-attn --no-build-isolation
# flash-attn>=2.5.0

# Additional dependencies for specific models
accelerate>=0.20.0
sentencepiece>=0.1.99
einops>=0.7.0
timm>=0.9.0
